// src/common/enums/Permission.ts
export enum Permission {
   adminRead = 'Admin.Read',
   adminWrite = 'Admin.Write',
   marketerAdmin = 'Marketer.Admin',
   marketerRead = 'Marketer.Read',
   marketerWrite = 'Marketer.Write',
   read = 'Read',
   scientistAdmin = 'Scientist.Admin',
   scientistRead = 'Scientist.Read',
   scientistWrite = 'Scientist.Write',
   write = 'Write',
}